LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-transition-flip-expected.txt


PASS Transition to a flipped state
FAIL Transition to an unflipped state assert_equals: expected 275 but got 300
